Extract RPC code to separate package
