+++ latte/src/main/java/gg/beemo/latte/broker/Exceptions.kt
@@ -1,5 +1,7 @@
 package gg.beemo.latte.broker
 
+import gg.beemo.latte.broker.rpc.RpcStatus
+
 sealed class BrokerException(message: String?) : Exception(message)
 class RpcRequestTimeout(message: String) : BrokerException(message)
 class IgnoreRpcRequest : BrokerException("Ignoring RPC request")
