+++ vanilla/src/main/java/gg/beemo/vanilla/KafkaRatelimitClient.kt
@@ -3,7 +3,7 @@ package gg.beemo.vanilla
 import gg.beemo.latte.broker.BrokerClient
 import gg.beemo.latte.broker.BrokerConnection
 import gg.beemo.latte.broker.IgnoreRpcRequest
-import gg.beemo.latte.broker.RpcStatus
+import gg.beemo.latte.broker.rpc.RpcStatus
 import gg.beemo.latte.logging.Log
 import gg.beemo.latte.ratelimit.SharedRatelimitData
 import gg.beemo.latte.util.SuspendingRatelimit
